break out credential definition into separate file
